Copied setup code from "OPT EDM Main PIC 2". The setup function itself, setupClock, and setupPortC are complete. Still need to finish setupPortA and setupPortB.
Added hardware definitions for the ports and pins. Changed code and added
banksel to each place writing or reading to/from latches/ports.

STILL NEED TO CHECK TO MAKE SURE BANKS ARE SET PROPERLY FOR OTHER CODE
NOW THAT IT IS BEING CHANGED FOR PORTS AND LATCHES.

Ran Source/Remove Trailing Spaces...bad idea as it made a lot of changes
to check in after removing spaces. This function should only be used
when there are no other changes to avoid clutter.

Signed-off-by: Mike Schoonover <mike@skoonie.com>