Various changes related to the EUSart -- these need to be copied to all other programs:
PCLATH now set before jumping to handleInterrupt in case that function is
in a different page.

Function handleSerialPacket has been changed to immediately copy the
packet length byte and release the receive buffer immediately so the next
packet can begin overwriting the buffer.

Calls to resetSerialPortRcvBuf removed from most instances as it is now
called by handleSerialPacket.

Functions parseCommandFromSerialPacket, handleLCDDataPacket, and
handleLCDInstructionPacket now use FSR1 instead of FSR0 so that some of
the functions called can modify FSR0 without conflict.

Function resetSerialPortRcvBuf now reads the serial port receive register
RCREG buffer twice to clear any pending interrupts.

Function handleInterrupt now checks for serial port interrupts when a
Timer0 interrupt is handled. This was mostly done so that the EUSART
simulator interrupts would be checked...the simulator seems not to trigger
interrupts for the receive when loading simulated data. It is probably
better for the regular operation as well.

Added function handleLCDBlockPacket -- only needed for this program and
does not need to be copied to others.

Signed-off-by: Mike Schoonover <mike@skoonie.com>