Added "high" to all movlp...required to work properly.
Used addfsr instead of incf for FSR0 in handleSerialPacket. Checksum
calculation was failing for large packets as the pointer was crossing
a 256 byte page boundary. The buffer is addressed as linear memory and can
be located anywhere.

Set the baud rate to 9,600. The opto-isolator H11N1 does not switch fast
enough for 57,600 -- the highs are 10uS instead of 17uS.

Capitalized C in I2C_COM_ERROR...variable is not used anywhere in the
code so no effect.

Added intScratch1 for debugging purposes.

Removed superfluous banksel INTCON -- it is a core register in all banks.

Signed-off-by: Mike Schoonover <mike@skoonie.com>